api_kafka/kafka/coinbase-fetcher/requirements.txt
requests
kafka-python
pyarrow
websocket-client